Reserve and release products

+++ src/main/kotlin/com/mc851/xcommerce/model/Product.kt
@@ -1,6 +1,6 @@
 package com.mc851.xcommerce.model
 
-data class Product(val id: Int, val name: String, val brand: String, val price: Int, val category: String?, val description: String, val imageUrl: String?)
+data class Product(val id: Long, val name: String, val brand: String, val price: Int, val category: String?, val description: String, val imageUrl: String?)
 
 data class Highlights(val highlights: List<Product>)
